requirements.txt
feedparser
psycopg[binary]
python-dotenv